comments

+++ src/ray/common/id.h
@@ -101,7 +101,7 @@ class JobID : public BaseID<JobID> {
  public:
   static constexpr int64_t kLength = 2;
 
-  static JobID FromInt(uint32_t value);
+  static JobID FromInt(uint16_t value);
 
   static size_t Size() { return kLength; }
